Chain "user_count()" off of "users()" too; leaves us with *one* implementation.
Downside is that the one implementation isn't going to be highly tuned SQL for
_every_ possible case.

Upside is that all three User-related methods now all accept the same
arguments, we know that they're all going to behave the same way, and its
going to be a whole lot easier to add support for other things in the future
(e.g. Groups) as we've only got to get it into _one_ underlying implementation
(instead of three separate ones).