Add filtering by account_id.
For workspace-based sources this pre-filters the workspaces by account_id.

For signals-based sources, we filter by the account_ids that the signal was
sent to.  If any of the accounts match the filter, the signal event is
included.

For person-based sources, if any shared account (between the viewer, actor,
and person) matches the filter, it is shown.  Person events aren't directly
associated with an account, so we can't filter on anything better than this.
If this doesn't match PMs expectations then we'll need a migration to
associate each person event with either the actor or person's primary account
then change this code to filter the same was as with signals.