"is_profile_hidden" is stored in the "users" table; make it accessible as a field on the User Homunculus.
This can help later on when we actually want to check whether or not a User
has their profile hidden as we don't have to go out into ST People and then
have it come back in and re-query the DB again for the User to pull up this
one field; we've already got it in memory, just use it.

Sucks donut goes to needing to include the default value from the DB in
`ST:U:Factory` so that a default gets set when new User records are created,
but we can eliminate that when we Moose-ify the User subsystem.