SQL migration to rebuild "account_user" and "user_account" VIEWs, with Group support.
The "account_user" VIEW looks the same as it did before, but now takes Group
memberships into consideration.  The "user_account" VIEW has been reworked
dramatically to be a far simpler VIEW/query than before, _but_ this meant that
`ST::User->ByAccountId()` had to be updated to match the new VIEW.

Functionally, this SQL update provides the same functionality that we had
before.  *HOWEVER*, as the number of User/Group/Workspace relationships grows,
these SQL queries degrade in performance *substantially*.  Consider this
commit a "make it work", _not_ a "make it fast".