These don't need to be DISTINCT queries.
Removing the DISTINCT qualification makes the queries faster, and also fixes
up some of our other unit tests which were expecting _multiple_ results.  I
hadn't realized that the code using this higher up the foodchain was able to
grok multiple/duplicate results, so I'm pleased to change it back to this for
a few reasons.