Whew.  1st Pass of Active Content is done.
This test creates two workpaces.  It creates pages in both workspaces.  It views lots of them a different number of times in order to predict which pages will appear on the widget (some should be cut off).

It watches a page, it emails a page, it edits pages.

It opens the dashboard and confirms all that for one workspace - and confirms that the things from a different workspace are absent.  Then it switches workspaces and does the opposite.  It does this while logging in as two different users.

I modified the output so it won't trip up a TAP parser (redirecting output to devnull) and created a setup page which can be included by other tests in order to generate some behavior.

I am considering creating a fake data account we can import, thus allowing us to test 1 week vs 1 month vs 1 year records - however, that has a problem as it will age out over time. I am open to ideas so I can test while I test.