Remove the "selected_only" parameter to `ST::WS::Roles->WorkspacesByUserId()`.
This "selected" feature was a left-over from S2 that has no use in S3.  Its
also a value that _isn't_ changable any more in the UI, and the DB sets all
new UWRs up to be "selected" automatically.  One day we may re-add this
feature, but for now its easier to remove it as its getting in the way (and
*isn't* be used visible anywhere except down here at the bottom of the
foodchain).