Don't assign Role in WS for "indirect" Users.
If during Workspace import we see that a User record is marked as being
"indirect", then *don't* give that User an explicit Role in the Workspace; his
membership is _indirect_ as is defined via some other means (e.g. Group
membership).

By default, all Users are treated as having "direct" membership in the
Workspace, preserving compatibility with exports from previous versions of ST.