Export "indirect" Users during Workspace export.
When exporting a Workspace, `ST::Workspace` will export information on all of
the Users that have a *direct* Role in the Workspace.  For Users that _only_
have an indirect or transitive Role in the Workspace, *we* are going to ensure
that the information for those Users gets properly exported.

On import, new/current versions of ST will automatically detect the "indirect"
flag and will recreate the User's membership in the Workspace via the Group.

When dealing with "export on new, import on old" scenarios, that version of ST
won't have any notion of "Groups" to begin with, and will do the best it can
by giving the User's an explicit "member" Role.  Hey, "infinite forwards
compatibility" isn't perfect, but IMHO this is better than "the Users that had
memberships via a Group were ignored entirely".