Keep the ST::Role cache up to date.
Otherwise we end up with it getting out of sync any time we update or remove a
Role.  Admittedly that doesn't happen very often, but when it does we'd be
stuck with the old Role sticking around until the process dies.

This commit also switches the caching over to use `ST::Cache` instead of just
using a hash.  Although `ST::Cache` is in some sense just overhead on top of a
hash (the current back-end implementation), having it abstracted out this way
(a) lets us replace that back-end to a shared-process cache if/when we want
to, and (b) ensures that it gets cleaned up automatically at the end of an
HTTP request.