Go back to per-process persistent cache.
Further discussion w/stash and we've come to the conclusion and agreement that
these things change so rarely that we don't have to worry about cache
coherency across processes or HTTP requests.  So, move this back to being an
in-process hash and get rid of the overhead that `ST::Cache` layers on top of
that.