Avoid some KS locking issue by forcing DESTROY.
I don't *think* this is masking a bug in KS, just fixing an improper usage in
this test.  What I think is happening is that DESTROY is trying to flush out
to the "segments" file and that holding open the searcher is causing the file
to stick around or stay locked.

In addition, the test count is updated for the "ceqlotron runs all jobs"
change.