Refactor LDAP filter building into ST::LDAP
This used to live lower-down in the code but we now have a need for more
general LDAP Filter generation, in a place which can be re-used by both the
LDAP "User" and "Group" factories.

Unit tests updated to remove test cases where we were explicitly looking to
make sure that this global LDAP filter was being applied by the lower-level.