New "group_filter" and "group_attr_map" LDAP cfg.
Adds a *separate* filter and LDAP attribute map for _Groups_, so that its
distinctly different than the one used for _Users_.

The LDAP Group Attribute Map is _optional_, on purpose; existing installs that
upgrade aren't going to have one in place, and we're not going to be able to
guess one correctly to build a migration that sets it up automatically.

So... allow for this Group Attr Map to be missing from the LDAP config, *but*
if we've got one then it has to contain all of the basic required attributes
in it.