"refresh" command, to refresh *both* Users+Groups.
Short-hand helper method, so that admins can more easily refresh *all* of
their LDAP data in one fell swoop.

Note that we refresh Users first _on purpose_, so that when we get around to
refreshing the Groups we already have up-to-date information on all of the
Users and we're going to have a higher hit rate on our local DB cache.