Fix {bz: 2923} by allowing the display of "last cached" data from LDAP.
When we have a case where we have a User that *used to* exist but that we
can't find in LDAP any longer, return a "Deleted User" homunculus that allows
for the display of the "last known good data for that User".

The User record *IS* still considered a "Deleted User" (as it still has a
ST::User::Deleted homunculus), but the last cached first/last/email is
available for display.