Pass through more values to "Deleted User" objects when they go missing.
When we instantiate a User object and (for some reason) we're unable to find
them (e.g. "LDAP server down/missing", "User completely relocated in LDAP and
can no longer be found"), pass through all of the "last known, cached data"
for the User, *rather than* only a limited subset of said data.