Allow for a "debug_level" to be set for OpenLDAP.
Sometimes (e.g. when doing perf testing) you may want to fire up OpenLDAP with
a custom debugging level.  Make this possible by allowing for a `debug_level`
option to be passed through to OpenLDAP when bootstrapping it.

For perf testing this *can* have a noticable impact, as with the default debug
level of "256" OpenLDAP can spend a great deal of time firehosing stuff to the
logs (which in turn slows down your perf tests).