In-memory cache of Groups (by Group Id).
The most common lookup done for Groups is by "group_id", so allow for these
lookups to be cached.  When updating the memberships for Groups that have a
large number of Users, this is a *HUGE* win as we don't have to continually
dip back into the DB and re-vivify the Group object again for each
UserGroupRole that is created/updated/removed.