LDAP Groups benchmark, for 20k Users, +- 5000 User churn.
Benchmark test to vivify a 25,000 User Group into ST, add 5000 Users, remove
5000 Users, then re-vivify the Group.

Test focusses on "how long does it take to do this?" not on the low-level
correctness of the results (which are tested elsewhere in other unit tests).

Test takes a *long* time to run, and is dominated by either "time spent in
OpenLDAP getting the results", or by "time spent parsing the results from
OpenLDAP down in the innards of Net::LDAP".

Bright side is that ST isn't the bottleneck in doing a 25k User feed, but the
down-side is that it also means that we don't have a lot of room for
improvement unless we either completely re-architect how we feed the Users in,
or we dig down into the Net::LDAP innards and punt patches upstream to help
LDAP perf in general.