Create/teardown GARs on Group creation and WS role assignment.
We do similar already for UARs when a User is created and assigned a Role in a
WS, so make sure that we do this for Groups too.  Most importantly, this
automatically populates the group_account_role table with the Primary Account
for a Group when the Group is created.