Allow configurable Role on all Groups "add" cmds
Instead of just relying on the default Role, at least make it possible for the
test writer to pass through their own specific Role Name when they're creating
Groups relationships.

If no Role is provided we _will_ use a default Role, but the author now has
the option of specifying their own specific Role when creating the test.