Optimize search for shortest-path to failing tests
Instead of doing a linear search through the set of tests that contributed to
the failure, do an optimized breadth-first traversal on a bisection of the
tests.

This not only allows us to discount and throw out whole groups of tests at a
single time, but going breadth-first lets us do that for larger groups of
tests earlier on in the game, increasing the chance that we're zeroing in
quicker than we would otherwise.