Only count failures from the test we're checking.
If _other_ unit tests failed along the way while we were checking for failures
in this unit test, don't count them.  This helps limit the focus to "just
those tests that contributed to the failure of *this* unit test", not ones
that contributed to the failure of _any_ of the tests.