Allow the candidate tests to be listed on cmd line.
If you already have a short-list of tests that you believe are the candidates
for failure, you can now list them on the command line (in order) rather than
letting "why-failing" try to calculate the list for you.

To specify multiple candidate tests, use this option multiple times.

e.g.
  why-failing -c t/one.t -c t/two.t -c t/three.t t/failing-test.t