Clear in-memory caches after fixture creation.
Creating the fixtures involves setting up an initial data set into the DB, so
flush the caches to ensure that this data is what's stored in the cache.

This fix cleans up a few unit-tests that fail on a clean test environment, but
which would run fine if _any_ test had been run before them (including
themselves).