This test stomps on the test-env; we're destructive.
This test also doesn't lend itself well to being refactored to use
"create_test_hub()" so that we've got a custom Hub/WS/User for this one test;
it has data files that live _outside_ of the test that contain information
specific to the WS+User that are used (and which are shared across other
tests, so we can't update them just for this one test).

We're taking the quick route and marking it as destructive.