Let LDAP Groups get created in a specified Account.
Up until this point we haven't needed a case for loading/creating an LDAP
Group into a specified Account in a single step.  This code lets us pass
through a "primary_account_id" when fetching the Group, and if we end up
vivifying a new Group object it'll end up in that Account.