Some more destructive tests.
Each of these tests creates new pages in the "admin" (with extra pages) WS,
leaving it dirty for the next test.

Looking at the tests, it looks like they actually _need_ the materials from
the "extra pages" in order to function properly, so for now I'm just marking
the tests as destructive so they don't pollute things for downstream tests.