Fixture reduction; use custom Hub+WS+User, not "admin".
Rather than using the "admin" WS and then creating new content into that which
we'd like to have indexed+searched, reduce the fixtures down so that we're
using our own _custom_ Hub+WS+User.

Saves time not only from having to bootstrap the "admin" WS when we have a
fixture conflict, but the custom Hub+WS+User are cleaned up automatically (so
we're not polluting the system for downstream tests).