Fixture reduction; only needed "admin" so we could create a test Acct.
We've since added a method to create a test account but bypassing the
factories during its (the Accounts) creation, thus eliminating the need for a
Hub (which is what we needed the Admin WS for here; to make sure that we had
the default test user record created in the DB).

By switching to use this other Account creation mechanism, we eliminate the
need to have the "admin" WS created in the first place, so we can reduce our
fixture requirements.