Fixture reduction; custom Hub+WS+User instead of stomping on Admin WS.
Only needed the Admin WS here so that we could create some new pages in it and
make sure that the backlinks *contained* the new things that we were adding.

We could just as easily do that in any other WS, which allows us to reduce the
fixtures by creating our own custom WS for the test.