Don't need "clean" here any more; we auto-cleanup at end of test.
Test was originally created _before_ we added the auto-cleanup materials to
the test fixtures, and as a result it needed to know that the target for the
rename did not yet exist.

Now that we clean these things up automatically at the end of each test,
though, this isn't necessary any longer.  We still need the "help" WS to be
present (so default pages can get copied over into the WS), but we don't need
to clean things out first.