Implement "list-groups", using *our* LDAP stack.
Makes it possible for the admin to run `st-ldap list-groups` and get a list of
*all* of the Groups that we can find/see in their configured LDAP Group
Factories.

This is *extra useful* because it uses *OUR* LDAP stack, and *not*
"ldapsearch"; when Support runs this command they know that they're seeing
everything that ther rest of our system should (as opposed to using
"ldapsearch" and still being left with the question of "if I can see it via
ldapsearch why can't ST find it?").