Run TheCeq jobs synchronously in wikiDtests too.
Note, though, that some of the wikiDtests suffer from "per-second granularity"
and as a result we need to sleep 1s before running the queue in order for the
tests to run cleanly.

When we shelled out this was generally the case as it'd take ~1-3s for the
process to fire up, but this also was the cause of repeated spurious failures.
If this sleep here fixes the issue, we should consider moving it over into the
actual wikiDtest and getting it out of the fixture (as only *some* Jobs are
susceptible to it, so only those test should incur the overhead of having to
do the sleep).