Pass through the default/fallback NTLM Domains.
Its possible for MSIE to make an NTLM SSO handshake and *not* provide us with
the name of the NTLM Domain that we're supposed to authenticate the User
against.

Be sure to pass through the configured "Default" and "Fallback" NTLM Domains,
so that `Apache::AuthenNTLM` has an idea what Domains it should attempt to
perform authentication for/against when this happens.