Genericize determination of entity/collection we're operating on.
We're starting to add in more combinations for things like "add-member" and
"remove-member", so we need to better genericize the way that we check to see
*which* of the combinations we're dealing with.

"_type_of_entity_collection_operation()" takes a list of possible
entity/collection combinations (e.g. "user-workspace"), and checks the
available CLI args to see which combination can be satisfied (returning the
first satisfied combination, so order *is* important).  If none of the
requested combinations can be satisfied, this method throws an error
explaining to the User that they need to specify "one of x,y,z *AND* one of
a,b,c".