Method to determine what Test Directory to run in.
Normally this is "t/tmp/", but if we're running multiple tests in parallel
this can be pushed down one more level via a "HARNESS_JOB_NUMBER" environment
variable (e.g. "t/tmp/1/", "t/tmp/2/", etc).

All sorts of other things are going to need to be able to determine which slot
or directory they're supposed to be running in, so provide some methods to
make this discernable.