Use 'test_dir()' to figure out where tests run from.
Also, we _don't_ need to abs_path() the resulting directory... it may not
exist yet (e.g. we're figuring out which directory to use for running the
tests under, *before* we've actually bootstrapped any fixtures that create
that directory).

Fortunately, "_user_checkout_dir()" has *already* been turned into an absolute
path, so if we just tack the test directory onto the end of that, we should be
ok.