Run TheCeq in-process in this unit-test.
No need to shell out to run this; we're not testing TheCeq, we're testing the
search results.  TheCeq is flushed so that we know that we've got all of the
Workspaces indexed when we do the search.

Doing this in-process should speed the test up noticably (which is good,
because this is currently the _slowest_ unit test we have).