Take test slot into consideration when calculating LDAP port number.
This prevents a race condition in the LDAP tests, when they're run in parallel.

When run in parallel, multiple tests try to start up OpenLDAP on the same port
at the same time.  One succeeds, but the rest see an LDAP server running where
they wanted it to be (and thus incorrectly think that its _their_ LDAP
instance).