Explicitly note that WS membership makes you an "Affiliate" w.r.t. the Account.
Although this is the default UserAccountRole right now, it is expected that
this will change at some point in the future (so that we can allow Users to be
a "Member" in their Primary Account).

To save ourselves some headache when we get there, explicitly mark these as
being an "Affiliate" Role so we don't have to update them when we get there.