Method to help query the Roles a User has in an Account.
A User can have a Role in an Account for a *variety* of different reasons, and
we want to be able to ask the question "what Role(s) does the User have?"
without caring where they come from or how he got them.

This helper method (and its package) are in the same vein as
ST::Workspace::Roles; capture all of the work for working on the
Roles/relationships that a User+Account can have with one another in one
place.

Unlike Users+Workspaces, though, Accounts have an *extra* level of indirection
that they offer; a User could have access to a Workspace in an Account (either
directly or through a Group) and that Group+Workspace relationship grants the
User a basic "Affiliate" Role in the Account.