Handle UserAccountRole upgrades/downgrades.
*WHEN* we eventually support Roles for a User in an Account other than
"Affiliate", this code will automatically upgrade/downgrade their Role in the
Account.

This commit mirrors work that was done for GroupAccountRoles, but for Users.

NOTE: The reason we *don't* yet support a UAR other than "Affiliate" is that
it changes the behaviour of the User+Account relationship (and to this point
we're still preserving the old behaviour, while building up code that'll let
us change it in the near future).

We're not there yet, but this code keeps moving us in that direction.  When we
get there, we are aware that we're going to need a data migration to update
the UARs for Users so that they're at least a "Member" in their Primary
Account, but we'll cross that bridge when we get there.