Groups re-import into their *original* Primary Account, on Account import.
When exporting Group information in an Account export, include the name of the
Group's Primary Account.  On Account import, the Group will be re-imported
back into that Account as its Primary Account.

If the Group's Primary Account doesn't exist on this ST instance, a new
Account with the same name is created, and the Group is assigned that Account
as its Primary Account.

This change required a major rewrite of one of the existing export/import
tests that dealt with merging of Group membership lists when an existing Group
was found.  Now that we're re-creating the original Account that the Group was
in, the test wasn't detecting "Group exists" any more.  Test was rewritten so
that it now utilized _two_ Accounts (one which was left alone, the other which
was exported/imported).