Placeholder test for direct UAR preservation on Account export/import.
We don't (yet) support the ability to add a User to an Account directly other
than by changing their Primary Account.

We know its coming, though, so when we get to that point lets make sure that
we've got a test that fails here to highlight that we better make sure that
this is getting exported/imported properly with the Account.