Pass *ALL* of the WS data to hooks on import.
This matches up the behaviour with what we have on Workspace export; where
hooks/plugins are given the entire data structure and are allowed to add their
own data to said structure.

On import, be sure to pass the hooks/plugins the whole data structure again,
so that they can then extract the data that they left there on export.