Account export dumps *all* Users in the Account, and they go back to their original Primary Account on import.
Changing the way that Account export/import works...

Previously, we'd export "just the Users that have this Account as their
Primary Account".  This is being changed so that we now export "*all* of the
Users that have _some_ Role in or access to the Account".

On re-import, though, we don't want to add _all_ of those Users to this
Account as their Primary Account (as that may grant them added privileges or
visibility that they didn't have before).  So, on import, be sure to put them
back into their original Primary Account.  If that Account doesn't exist, go
re-create an empty Account with that name and put the User there.