Parse pagelinks in a separate process.
We are seeing occaisonal core dumps from the WikiText parser.
Rather than have these core dumps kill the entire job for this workspace,
we should catch & log the core dumps.  The links on this page will not
be processed, but the other pages in the workspace will be.

Yours Truly,
Kevin and Luke