New "Placeholder" Account Type; an Account which is _just_ a placeholder and which *can* be over-written on Account import.
Recent changes to the Account export/import model left us in a scenario where
we could have multiple Accounts which cross-reference one another, and that
the import behaviour set us up for *only* being able to import a single one of
the Accounts...

    Importing "Account A" would create empty/placholder Accounts for any other
    Account necessary to satisfy a Primary Account relationship for a User or
    Group.  However, this then meant that when we wanted to import "Account B"
    it would fail as the Account already existed within the system.
    Incomplete, but it existed.

    Using "st-admin account-import --overwrite" was also sub-optimal as that
    completely erased "Account B", which cascades through to things living in
    that Account which may have actually had a Role or relationship with
    something back in "Account A".  Using "--overwrite" would allow you to
    import "Account B", but in such a manner that it broke all the
    relationships that its Users or Groups had to things living over in
    "Account A".

Thus, the "Placeholder" Account type.... "I'm an Account, but I'm explicitly
stating that I'm *just* a placeholder and that I *can* be over-written during
Account import".

When importing the second/third/etc Account onto the system, the Account
import will see that these other Accounts were created for the purpose of
being a placeholder and they'll (a) fill in the missing details, and (b) set
the Account Type to something reasonable.

Yes, I agree that this is *sub-optimal* both in design and in implementation.
Unfortunately, I don't believe that it is something that can be addressed
properly without significantly re-thinking how we do export/import of multiple
Accounts, potentially making *drastic* changes to the export format.