Tell prove to save its state after the test run.
By having "prove" save its state, you've now got a history of the order in
which the tests were run.  This makes it possible to re-run tests in the same
order, or to be able to ask prove to do things like "only run me the tests
that failed recently".

When working in a dev-env to try to fix up failures before pushing upstream,
this can be a great workflow:

 1. "make test"
 2. see failing tests, make fixes
 3. "st-prove --state=failed,save"
 4. repeat 2+3 until you're out of failures
 5. profit!