Do not remove a user's membership from their old primary account.
Previously, a user could only be a member ( read "directly associated" ) with
a single account. All other relationships to the account were through
workspaces or groups. Since it is now possible for a user to have membership
in multiple accounts, we don't need to remove, and in fact don't _want_ to
remove them, when we assign a new primary account.

This is the basic fix for {bz 3326}, but it also opens up a _deep_ rathole.