Update ST::User->deactivate(), implement ->reactivate().
When we deactivate a user, we want to be sure to remove them from any
accounts other than their primary that thay may have also had a membership in.

When we reactivate a user, we want to be absolutely sure that we remove then
from the Deleted account.

When in this code, I noticed that deactivating a user does _not_ pull them
from their groups. After some discussion, I've left that code out until we
have a spec for doing so.