Fix test directory determination so it works in a VE.
Ken had noted that this wasn't working properly for him in a QA VE, as we'd be
taking the abs_path() of something that we hadn't necessarily created yet (the
tmp directory).

Switched it around a bit so that the detected "nlw_dir" is *always* turned
into an absolute path, and then everywhere after we only just have to tack on
the bits that we need at that time. Being that "nlw_dir" *HAD* to exist in the
first place (as otherwise we wouldn't have detected it), taking the abs_path()
here is safe and should provide consistent results.

I've tested this (a) in a dev-env, (b) w/fdefs, and (c) by copying it into a
VE that was upgraded from a previous release.  Worked fine in all occasions.