Convert user-sets from GiST to B-tree indexing.
The primary reason for this change is that the migration was taking literally
hours to generate the GiST index on `user_set_path.vlist`.  The actual time it
would have taken isn't known; we stopped it at 6 hours.

We still use the intarray module for the `idx`, `sort`, `uniq` and `icount`
functions.

Removes the 'via_set_id' column from user_set_path; we can still use the vlist
column to obtain that info.  This affects the user_set_include_tc view which
will now return distinct role_ids and not role_id,via_set_id pairs.