Automatically assign Users into their Primary Account, when we create their UserMetadata record.
At the point of creating the UserMetadata record, we *know* that we're (a)
creating a new User, and (b) assigning them into some Primary Account.  So...
here's a good place to have that actual assignment take place (and to ensure
that the User gets properly added to the AUW for that Account).