Helper methods, to make sure that no warnings/errors were logged.
At times we're not really concerned specifically about what was logged, but
only that there were no errors or warnings logged that would indicate a
problem.

If it was successful we may have logged something informational or for
debugging, but everything ran ok.  In the event of a problem, we would've
logged an error somewhere along the way and we can watch for those more easily
now.