Don't use ST::Account in migration 55.
Because things are in such flux with the user-set migration, let's keep this
as simple as possible and require nothing but a dbh.

Let's also defensively select around any accounts that may already have the
plugin installed so we don't have to worry about failing trying to create
duplicate records.