Apply the UserSetContained Role to users/groups.
Allows for the use of the UserSetPerspective code.

The included test is the primary coverage for UserSetPerspective and
UserSetContained, at this point in time.

Something we don't need quite yet, but will still be useful, is
sorted_supergroup_roles - a way to report on a group's roles inside of other
supergroups.  Sent to the patches directory.