TC: BCP Groups in Workspace.  Reflects changes in business process.
Now when I create a group, my user is set as the admnistrator.  So when this code calls:

create_group foo
add_member joe@schmoe.com

Group foo now has TWO members in it - %%username$$ AND joe@schmoe.com.  So I needed to change
the test to expect 2 users. FTW.