Display errors from "st-admin" as test diagnostics.
Luke had made a change recently to get STDERR displayed as part of the test
run, so that if there were any errors when running st-admin that they'd be
visible (in case they were *unexpected* errors).  Unfortunately, some of our
tests relied on the capture of STDERR so that it could be verified for
correctness (and those tests broke as a result).

So... lets capture *both* STDOUT and STDERR, check their combined values for
verification, and _also_ output STDERR as test diagnostics.  That'll get tests
working again, but also make any warnings/errors obvious in case they're
unexpected.

To do this, though, involves a bit of a refactor and cleanup to how we're
doing both the "in process" and the "shelled out" calls to st-admin, so that
we can properly capture STDOUT/STDERR separately (but simultaneously).

Further, regardless of whether its run "in process" or "shelled out", the
visual display of the results should now be *identical*; previously if you
shelled out then STDERR was just "warn"ed, while its now issued as a test
diagnostic (so as not to disrupt the TAP stream).