Begin implementing default permissions matrix for groups.
This will have to serve until we get ACL's going for groups. Until that time,
we should avoid using role-based assumptions about permissions (eg. "This user
is an admin, therefore they should be able to do XXX) and consult our
permissions matrix. When we do get around to implementing ACL's, we'll only
need to change this sub and everything should 'just work'.