Ensure that the required data exists, even if we're not doing a schema update.
This allows for us to add in new "required data" and to have it automatically
populated into an appliance on update, *even if* there is no other DB
migration that needs to be done.

FYI... "add required data" should be a no-op if all of the required data
already exists.