Fixture reduction; these tests don't need a full "admin" Workspace.
Instead, they can make do with an Admin Workspace that has *no* pages in it,
allowing us to cut down on initial fixture setup.

These tests are also *good* candidates for further reduction to actually
eliminate the need for Workspace setup entirely, and just have the tests
create their owk test Workspaces as needed (so they only rely on the "db"
fixture, and then do the rest of the setup themselves).  Or, perhaps we bring
back an "empty" Workspace; an arbitrary Workspace that has *no* pages in it
whatsoever.