Fixture switch; "admin_no_pages" => "empty".
Makes it more obvious that we don't need any of the "Admin" facets of the
Workspace for the test, and that all we really need is just an "empty"
Workspace to work against.

Also allows us to make better use of the fixture cache as we won't conflict
against other tests that need one of the various "admin" or "foobar" fixtures
(which can conflict with one another).