Cascade UPDATEs through as well as DELETEs.
When migrating an "Account" to a "Group", the underlying "user_set_id" is
changed in the "signal_user_set" table.  In order to ensure that we continue
to maintain referential integrity, we need to make sure that those raw UPDATEs
get properly propogated through to the "recent_signal_user_set" table.

This case was highlighted by a failing `t/ST/Acct/Transformer-signals.t` unit
test.