Make $UPLOAD_DIR safer.
There seems to be a problem in tests where our upload dir is not created with
global read/write/execute perms. This isn't a problem in production because
the same user, www-data, will always be the one operating on the dir. In
tests, however, there's any number of users that will need access, so let's
just use our generic temp dir and be a bit more crafty with uploaded file
names.