Push initial Group relationship creation down into the Factory.  Fixes {bz: 3498}
When a new Group record is created (whether that's through explicit creation,
or through auto-vivification of an LDAP Group), we need to set up an initial
set of relationships for the Group.

Bug {bz: 3498} identified that this _wasn't_ happening properly for LDAP
Groups; they were being created but not properly assigned a Role in their
Primary Account.

It was, however, happening properly for Groups that were being explicitly
created.  This commit pushes that "create the initial relationships" code down
into the base Group Factory class and ties it in so that it now happens
properly for both "Default" and "LDAP" Groups.