Optimize reckless deletion of Accounts/Groups/Users, and ensure cache cleanliness.
When deleting an Account/Group/User, don't disconnect from the DB outright
(thus forcing the next DB operation to have to create a brand new connection
to the DB).  When doing cleanup at the end of a unit test, this would cause us
to have to reconnect to the DB for each and every User/Group/Account that we
were cleaning up.

Also, ensure that the caches are cleared when we delete something recklessly;
it might have been cached (either as an object _or_ in a count or summary) by
something else.