Remove last two vestiges of "watchlist_dashboard_length".
Was an old S2 Dashboard option (and we're not on S2 any more).

All of the other code/preferences/templates relating to this were removed by
Ken back in July 2009.  I'd left a "note to self" to clean these last two
things up, and its been sitting in my Inbox ever since.