Defensive coding: ensure we don't recurse while doing LDAP Group refreshes.
Some customers recently experienced problems where LDAP Group refreshes were
taking too long and causing undue load on both their LDAP directories and
their Socialtext Appliances.

Turned out that we'd accidentally introduced a codepath that caused LDAP Group
Refresh operations to recurse; while refreshing the Group we'd trigger some
other action that in turned tried to instantiate the same LDAP Group object
(which in turn then tried to refresh the Group).

The codepath in question was recently refactored out as part of the user-set
work, but I felt that it was worth committing this code to ensure that we
don't accidentally regress again in the future.

@gtermars has verified that this code catches a recursive LDAP Group lookup
when applied to "release-3.7.0.11-1", and that the recursion does *not* occur
on current "master".

FYI, the test case in `t/live-ldap/Socialtext/LDAP/Operations-RefreshGroup.t`
*is* sufficient to trigger the recursion that was seen in "release-3.7.0.11-1"
(and as such, no additional test case has been added here).