Delete Groups recklessly in wiki tests. TC: Groups
When we delete Groups in a wiki test, its ok for us to be a bit
reckless/destructive in how we do that.

Going this route allows us to create an LDAP Group and then delete it later on
(as you can't otherwise delete/purge an LDAP Group from ST _yet_).